make config option quote default to "single"

+++ src/config/defaults.ts
@@ -5,6 +5,6 @@ export const defaults: Required<
 > = {
   outputDir: '.features-gen',
   verbose: false,
-  quotes: 'double',
+  quotes: 'single',
   language: 'en',
 };
